requirements.txt
asgiref==3.10.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-database-url==3.1.2
Django==5.2.7
django-allauth==65.13.0
django-cors-headers==4.9.0
django-redis==6.0.0
django-schema-viewer==0.5.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
elastic-transport==8.17.1
elasticsearch==8.19.2
elasticsearch-dsl==8.18.0
google-auth==2.43.0
gunicorn==25.1.0
idna==3.11
packaging==26.0
pillow==12.0.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
razorpay==2.0.0
redis==7.0.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.12.0
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
